+++ frontend/src/app/(routes)/cosmwasm/components/AllContracts.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const AllContracts = () => {
-  return <div>AllContracts</div>;
+  return <div className="h-1/2 flex-center-center">Comming Soon...</div>;
 };
 
 export default AllContracts;
