+++ frontend/src/app/(routes)/cosmwasm/components/ContractNotSelected.tsx
@@ -1,0 +1,18 @@
+import Image from 'next/image';
+import React from 'react';
+
+const ContractNotSelected = () => {
+  return (
+    <div className="flex items-center gap-2">
+      <Image
+        src="/file-upload-icon.svg"
+        width={32}
+        height={32}
+        alt="Search Contract"
+      />
+      <div>Select or search contract</div>
+    </div>
+  );
+};
+
+export default ContractNotSelected;
