+++ frontend/src/app/(routes)/cosmwasm/components/ProvideFundsJson.tsx
@@ -1,5 +1,6 @@
 import { TextField } from '@mui/material';
 import React from 'react';
+import { queryInputStyles } from '../styles';
 
 interface ProvideFundsJsonI {
   fundsInput: string;
@@ -25,8 +26,8 @@ const ProvideFundsJson = (props: ProvideFundsJsonI) => {
   };
 
   return (
-    <div className="bg-[#FFFFFF0D] p-4 rounded-2xl">
-      <div className=" border-[1px] border-[#ffffff1e] hover:border-[#ffffff50] rounded-2xl relative">
+    <div className="provide-funds-input-wrapper">
+      <div className="provide-funds-input">
         <TextField
           value={fundsInput}
           onChange={handleFundsChange}
@@ -42,31 +43,9 @@ const ProvideFundsJson = (props: ProvideFundsJsonI) => {
               },
             },
           }}
-          sx={{
-            '& .MuiTypography-body1': {
-              color: 'white',
-              fontSize: '12px',
-              fontWeight: 200,
-            },
-            '& .MuiOutlinedInput-notchedOutline': {
-              border: 'none',
-            },
-            '& .MuiOutlinedInput-root': {
-              border: 'none',
-              borderRadius: '16px',
-              color: 'white',
-            },
-            '& .Mui-focused': {
-              border: 'none',
-              borderRadius: '16px',
-            },
-          }}
+          sx={queryInputStyles}
         />
-        <button
-          type="button"
-          onClick={formatJSON}
-          className="border-[1px] border-[#FFFFFF33] rounded-full p-2 text-[12px] font-extralight top-4 right-4 absolute"
-        >
+        <button type="button" onClick={formatJSON} className="format-json-btn">
           Format JSON
         </button>
       </div>
