+++ frontend/src/app/(routes)/cosmwasm/components/QueryContract.tsx
@@ -5,6 +5,7 @@ import { TxStatus } from '@/types/enums';
 import { CircularProgress, TextField } from '@mui/material';
 import React, { useEffect, useState } from 'react';
 import { queryInputStyles } from '../styles';
+import { setError } from '@/store/features/common/commonSlice';
 
 interface QueryContractI {
   address: string;
@@ -14,9 +15,19 @@ interface QueryContractI {
 
 const QueryContract = (props: QueryContractI) => {
   const { address, baseURLs, chainID } = props;
+
+  // ------------------------------------------//
+  // ---------------DEPENDENCIES---------------//
+  // ------------------------------------------//
   const dispatch = useAppDispatch();
-  const { getContractMessages, getQueryContractOutput } = useContracts();
+  const { getContractMessages, getQueryContract } = useContracts();
+
+  // ------------------------------------------//
+  // ------------------STATES------------------//
+  // ------------------------------------------//
+  const [queryText, setQueryText] = useState('');
   const [contractMessages, setContractMessages] = useState<string[]>([]);
+
   const queryOutput = useAppSelector(
     (state) => state.cosmwasm.chains?.[chainID]?.query.queryOutput
   );
@@ -24,16 +35,9 @@ const QueryContract = (props: QueryContractI) => {
     (state) => state.cosmwasm.chains?.[chainID]?.query.status
   );
 
-  const [queryText, setQueryText] = useState('');
-
-  useEffect(() => {
-    const fetchMessages = async () => {
-      const { messages } = await getContractMessages({ address, baseURLs });
-      setContractMessages(messages);
-    };
-    fetchMessages();
-  }, [address]);
-
+  // ------------------------------------------------//
+  // -----------------CHANGE HANDLERS----------------//
+  // ------------------------------------------------//
   const handleQueryChange = (
     e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
   ) => {
@@ -44,36 +48,85 @@ const QueryContract = (props: QueryContractI) => {
     setQueryText(`{\n\t"${msg}": {}\n}`);
   };
 
+  const formatJSON = () => {
+    try {
+      const parsed = JSON.parse(queryText);
+      const formattedJSON = JSON.stringify(parsed, undefined, 4);
+      setQueryText(formattedJSON);
+      return true;
+      /* eslint-disable @typescript-eslint/no-explicit-any */
+    } catch (error: any) {
+      dispatch(
+        setError({
+          type: 'error',
+          message: 'Invalid JSON input: ' + (error?.message || ''),
+        })
+      );
+    }
+    return false;
+  };
+
+  // --------------------------------------//
+  // -----------------QUERY----------------//
+  // --------------------------------------//
   const onQuery = () => {
+    if (!queryText?.length) {
+      dispatch(
+        setError({
+          type: 'error',
+          message: 'Please enter query message',
+        })
+      );
+      return;
+    }
+    if (!formatJSON()) {
+      return;
+    }
+
     dispatch(
       queryContractInfo({
         address,
         baseURLs,
         queryData: queryText,
         chainID,
-        getQueryContractOutput,
+        getQueryContract,
       })
     );
   };
 
+  // ------------------------------------------//
+  // ---------------SIDE EFFECT----------------//
+  // ------------------------------------------//
+  useEffect(() => {
+    const fetchMessages = async () => {
+      const { messages } = await getContractMessages({ address, baseURLs });
+      setContractMessages(messages);
+    };
+    fetchMessages();
+  }, [address]);
+
   return (
     <div className="flex gap-10">
-      <div className="query-field flex flex-col gap-4">
-        <div className="flex gap-4 flex-wrap">
-          {contractMessages?.map((msg) => (
-            <div
-              onClick={() => handleSelectMessage(msg)}
-              key={msg}
-              className="px-4 py-2 rounded-2xl bg-[#FFFFFF14] cursor-pointer hover:bg-[#ffffff26]"
-            >
-              {msg}
-            </div>
-          ))}
+      <div className="query-input-wrapper">
+        <div className="space-y-4">
+          <div className="font-medium">Suggested Messages:</div>
+          <div className="flex gap-4 flex-wrap">
+            {contractMessages?.map((msg) => (
+              <div
+                onClick={() => handleSelectMessage(msg)}
+                key={msg}
+                className="query-shortcut-msg"
+              >
+                {msg}
+              </div>
+            ))}
+          </div>
         </div>
-        <div className="relative flex-1 border-[1px] rounded-2xl border-[#ffffff1e] hover:border-[#ffffff50]">
+        <div className="query-input">
           <TextField
             value={queryText}
             name="queryField"
+            placeholder={JSON.stringify({ test_query: {} }, undefined, 2)}
             onChange={handleQueryChange}
             fullWidth
             multiline
@@ -92,7 +145,7 @@ const QueryContract = (props: QueryContractI) => {
           <button
             onClick={onQuery}
             disabled={queryLoading === TxStatus.PENDING}
-            className="primary-gradient h-10 rounded-lg px-3 py-[6px] absolute bottom-6 right-6 min-w-[85px]"
+            className="primary-gradient query-btn"
           >
             {queryLoading === TxStatus.PENDING ? (
               <CircularProgress size={18} sx={{ color: 'white' }} />
@@ -100,10 +153,17 @@ const QueryContract = (props: QueryContractI) => {
               'Query'
             )}
           </button>
+          <button
+            type="button"
+            onClick={formatJSON}
+            className="format-json-btn"
+          >
+            Format JSON
+          </button>
         </div>
       </div>
       <div className="query-output-box">
-        <div className=" border-[1px] border-[#ffffff1e] h-full rounded-2xl p-2 overflow-x-scroll overflow-y-scroll">
+        <div className="qeury-output">
           <pre>{JSON.stringify(queryOutput, undefined, 2)}</pre>
         </div>
       </div>
