+++ frontend/src/app/(routes)/cosmwasm/components/SearchContracts.tsx
@@ -1,8 +1,8 @@
-import CommonCopy from '@/components/CommonCopy';
-import Image from 'next/image';
 import React, { useState } from 'react';
 import DialogSearchContract from './DialogSearchContract';
 import useGetChainInfo from '@/custom-hooks/useGetChainInfo';
+import ContractSelected from './ContractSelected';
+import ContractNotSelected from './ContractNotSelected';
 
 interface SearchContractsI {
   chainID: string;
@@ -12,17 +12,23 @@ interface SearchContractsI {
 
 const SearchContracts = (props: SearchContractsI) => {
   const { chainID, handleSelectContract, selectedContract } = props;
+
+  // DEPENDENCIES
+  const { getChainInfo } = useGetChainInfo();
+  const { restURLs } = getChainInfo(chainID);
+
+  // STATE
   const [searchDialogOpen, setSearchDialogOpen] = useState(false);
-  const hanldeClose = () => {
+
+  // CHANGE HANDLER
+  const handleClose = () => {
     setSearchDialogOpen(false);
   };
 
-  const { getChainInfo } = useGetChainInfo();
-  const { restURLs } = getChainInfo(chainID);
   return (
     <>
       <div
-        className={`px-4 py-6 border-[1px] border-[#FFFFFF80] rounded-2xl max-h-20 flex ${!selectedContract.address ? 'cursor-pointer' : 'cursor-default'}`}
+        className={`selected-contract ${!selectedContract.address ? 'cursor-pointer' : 'cursor-default'}`}
         onClick={() => {
           if (!selectedContract.address) setSearchDialogOpen(true);
         }}
@@ -38,7 +44,7 @@ const SearchContracts = (props: SearchContractsI) => {
       </div>
       <DialogSearchContract
         open={searchDialogOpen}
-        onClose={hanldeClose}
+        onClose={handleClose}
         restURLs={restURLs}
         chainID={chainID}
         handleSelectContract={handleSelectContract}
@@ -48,41 +54,3 @@ const SearchContracts = (props: SearchContractsI) => {
 };
 
 export default SearchContracts;
-
-const ContractNotSelected = () => {
-  return (
-    <div className="flex items-center gap-2">
-      <Image
-        src="/file-upload-icon.svg"
-        width={32}
-        height={32}
-        alt="Search Contract"
-      />
-      <div>Select or search contract</div>
-    </div>
-  );
-};
-
-const ContractSelected = ({
-  contract,
-  openSearchDialog,
-}: {
-  contract: { address: string; name: string };
-  openSearchDialog: () => void;
-}) => {
-  const { address, name } = contract;
-  return (
-    <div className="flex justify-between items-center w-full gap-10">
-      <div className="flex items-center gap-10">
-        <div className="text-[14px]">{name}</div>
-        <CommonCopy message={address} style="" plainIcon={true} />
-      </div>
-      <button
-        onClick={openSearchDialog}
-        className="primary-gradient change-btn"
-      >
-        Change Contract
-      </button>
-    </div>
-  );
-};
