+++ frontend/src/app/(routes)/cosmwasm/components/UploadContract.tsx
@@ -22,6 +22,7 @@ interface UploadContractI {
   walletAddress: string;
   restURLs: string[];
 }
+
 interface UploadContractInput {
   wasmFile?: File;
   permission: AccessType;
@@ -30,16 +31,42 @@ interface UploadContractInput {
 
 const UploadContract = (props: UploadContractI) => {
   const { chainID, walletAddress, restURLs } = props;
+
+  // ------------------------------------------//
+  // ---------------DEPENDENCIES---------------//
+  // ------------------------------------------//
   const dispatch = useAppDispatch();
   const { uploadContract } = useContracts();
+
+  // ------------------------------------------//
+  // -----------------STATES-------------------//
+  // ------------------------------------------//
+  const [uploadedFileName, setUploadedFileName] = useState<string>('');
+  const [accessType, setAccessType] = useState<AccessType>(3);
 
   const uploadContractStatus = useAppSelector(
     (state) => state.cosmwasm.chains?.[chainID]?.txUpload.status
   );
   const uploadContractLoading = uploadContractStatus === TxStatus.PENDING;
 
-  const [uploadedFileName, setUploadedFileName] = useState<string>('');
-  const [accessType, setAccessType] = useState<AccessType>(3);
+  // ------------------------------------------------//
+  // -----------------FORM HOOKS---------------------//
+  // ------------------------------------------------//
+  const { setValue, handleSubmit, getValues } = useForm<UploadContractInput>({
+    defaultValues: {
+      wasmFile: undefined,
+      permission: AccessType.ACCESS_TYPE_EVERYBODY,
+      allowedAddresses: [{ address: '' }],
+    },
+  });
+
+  // ------------------------------------------------//
+  // -----------------CHANGE HANDLERS----------------//
+  // ------------------------------------------------//
+  const handleAccessTypeChange = (event: SelectChangeEvent<AccessType>) => {
+    setAccessType(event.target.value as AccessType);
+    setValue('permission', event.target.value as AccessType);
+  };
 
   const resetUploadedFile = () => {
     const fileInputElement = document.getElementById(
@@ -49,19 +76,9 @@ const UploadContract = (props: UploadContractI) => {
     setValue('wasmFile', undefined);
   };
 
-  const { setValue, handleSubmit, getValues } = useForm<UploadContractInput>({
-    defaultValues: {
-      wasmFile: undefined,
-      permission: AccessType.ACCESS_TYPE_EVERYBODY,
-      allowedAddresses: [{ address: '' }],
-    },
-  });
-
-  const handleAccessTypeChange = (event: SelectChangeEvent<AccessType>) => {
-    setAccessType(event.target.value as AccessType);
-    setValue('permission', event.target.value as AccessType);
-  };
-
+  // ------------------------------------------//
+  // ---------------TRANSACTION----------------//
+  // ------------------------------------------//
   const onUpload = async (data: UploadContractInput) => {
     const wasmcode = getValues('wasmFile')?.arrayBuffer();
     if (wasmcode) {
@@ -157,12 +174,12 @@ const UploadContract = (props: UploadContractI) => {
         </div>
         <div className="w-1/2 flex flex-col gap-6">
           <div className="space-y-6">
-            <div className="border-b-[1px] border-[#ffffff1e] pb-4 space-y-2">
+            <div className="attach-funds-header">
               <div className="text-[18px] font-bold">
                 Select Instantiate Permission
               </div>
               {/* TODO: Update the dummy description */}
-              <div className="leading-[18px] text-[12px] font-extralight">
+              <div className="attach-funds-description">
                 Lorem ipsum dolor sit amet consectetur adipisicing elit.
                 Necessitatibus fuga consectetur reiciendis fugit suscipit ab.
               </div>
@@ -177,10 +194,10 @@ const UploadContract = (props: UploadContractI) => {
       </div>
       <button
         disabled={uploadContractLoading}
-        className="primary-gradient rounded-lg py-[6px] flex-1 w-full"
+        className="primary-gradient upload-btn"
       >
         {uploadContractLoading ? (
-          <CircularProgress size={16} sx={{ color: 'white' }} />
+          <CircularProgress size={18} sx={{ color: 'white' }} />
         ) : (
           <>Upload Contract</>
         )}
