+++ frontend/src/app/(routes)/cosmwasm/cosmwasm.css
@@ -1,5 +1,17 @@
+.tabs {
+  @apply flex gap-10 items-center justify-center border-b-[1px] border-[#ffffff1e] mt-6;
+}
+
 .menu-item {
   @apply cursor-pointer px-2 text-[18px] h-9 flex items-center pb-[14px] leading-[21.7px];
+}
+
+.selected-contract {
+  @apply px-4 py-6 border-[1px] border-[#FFFFFF80] rounded-2xl max-h-20 flex;
+}
+
+.styled-underlined-text {
+  @apply font-bold underline underline-offset-[3px] cursor-pointer;
 }
 
 .change-btn,
@@ -33,18 +45,22 @@
   @apply bg-[#171429] text-[12px] font-semibold px-3 py-[10px] flex justify-center items-center rounded-lg;
 }
 
-.query-field,
+.query-input-wrapper,
 .query-output-box {
-  @apply w-[50%] bg-[#FFFFFF0D] rounded-2xl min-h-[380px] p-6;
+  @apply w-[50%] bg-[#FFFFFF0D] rounded-2xl min-h-[380px] p-4  flex flex-col gap-4;
 }
 
-.execute-field,
+.execute-field-wrapper,
 .execute-output-box {
   @apply w-[50%] max-h-[440px] p-6;
 }
 
-.execute-field {
-  @apply bg-[#ffffff0d] rounded-2xl;
+.execute-field-wrapper {
+  @apply bg-[#ffffff0d] rounded-2xl p-4 flex flex-col gap-4;
+}
+
+.execute-input-field {
+  @apply relative flex-1 border-[1px] rounded-2xl border-[#ffffff1e] hover:border-[#ffffff50];
 }
 
 .execute-output-box {
@@ -69,4 +85,61 @@
 
 .contract-item {
   @apply flex gap-4 justify-between items-center px-3 py-3 pr-1 rounded-full cursor-pointer border-[1px] border-[#ffffff32] hover:bg-[#ffffff32];
+}
+
+.execute-btn {
+  @apply h-10 rounded-lg px-3 py-[6px] absolute bottom-6 right-6 min-w-[85px];
+}
+
+.format-json-btn {
+  @apply border-[1px] border-[#FFFFFF33] rounded-full p-2 text-[12px] font-extralight top-4 right-4 absolute hover:border-[#ffffff50];
+}
+
+.attach-funds-header {
+  @apply border-b-[1px] border-[#ffffff1e] pb-4 space-y-2;
+}
+
+.attach-funds-description {
+  @apply leading-[18px] text-[12px] font-extralight;
+}
+
+.instantiate-input-wrapper {
+  @apply bg-[#ffffff0d] p-4 rounded-2xl flex-1;
+}
+
+.instantiate-input {
+  @apply relative h-full border-[1px] rounded-2xl border-[#ffffff1e] hover:border-[#ffffff50];
+}
+
+.provide-funds-input-wrapper {
+  @apply bg-[#FFFFFF0D] p-4 rounded-2xl;
+}
+
+.provide-funds-input {
+  @apply border-[1px] border-[#ffffff1e] hover:border-[#ffffff50] rounded-2xl relative;
+}
+
+.query-btn {
+  @apply h-10 rounded-lg px-3 py-[6px] absolute bottom-6 right-6 min-w-[85px];
+}
+
+.query-input {
+  @apply relative flex-1 border-[1px] rounded-2xl border-[#ffffff1e] hover:border-[#ffffff50];
+}
+
+.qeury-output {
+  @apply border-[1px] border-[#ffffff1e] h-full rounded-2xl p-4 overflow-x-scroll overflow-y-scroll;
+}
+
+.query-shortcut-msg {
+  @apply px-4 py-2 rounded-2xl bg-[#FFFFFF14] cursor-pointer hover:bg-[#ffffff26];
+}
+
+.add-funds-btn {
+  @apply rounded-lg px-3 py-[6px] font-medium leading-[20px] text-[12px];
+}
+
+.instantiate-btn,
+.upload-btn {
+  @apply flex justify-center items-center text-[12px] font-medium px-3 py-[6px] rounded-lg h-10 w-full;
 }
